refactor: align package name to group id
