+++ src/main/java/org/extism/sdk/chicory/CachedAotMachineFactory.java
@@ -1,4 +1,4 @@
-package org.extism.chicory.sdk;
+package org.extism.sdk.chicory;
 
 import com.dylibso.chicory.experimental.aot.AotMachineFactory;
 import com.dylibso.chicory.runtime.Instance;
