+++ src/main/java/org/extism/sdk/chicory/ExtismFunctionException.java
@@ -1,4 +1,4 @@
-package org.extism.chicory.sdk;
+package org.extism.sdk.chicory;
 
 public class ExtismFunctionException extends ExtismException {
