+++ src/main/java/org/extism/sdk/chicory/ExtismTypeConversionException.java
@@ -1,4 +1,4 @@
-package org.extism.chicory.sdk;
+package org.extism.sdk.chicory;
 
 public class ExtismTypeConversionException extends ExtismException {
