+++ src/main/java/org/extism/sdk/chicory/ExtismValTypeList.java
@@ -1,4 +1,4 @@
-package org.extism.chicory.sdk;
+package org.extism.sdk.chicory;
 
 import com.dylibso.chicory.wasm.types.ValueType;
