+++ src/main/java/org/extism/sdk/chicory/Manifest.java
@@ -1,4 +1,4 @@
-package org.extism.chicory.sdk;
+package org.extism.sdk.chicory;
 
 import com.dylibso.chicory.wasi.WasiOptions;
