+++ src/test/java/org/extism/sdk/chicory/HttpTest.java
@@ -1,4 +1,4 @@
-package org.extism.chicory.sdk;
+package org.extism.sdk.chicory;
 
 import com.dylibso.chicory.log.SystemLogger;
 import jakarta.json.Json;
